% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmcm_joint_stan.R
\name{bmcm_joint_stan}
\alias{bmcm_joint_stan}
\title{Run stan mixture cure joint model}
\usage{
bmcm_joint_stan(
  input_data,
  model_os = "exp",
  model_pfs = "exp",
  tx_name = "IPILIMUMAB",
  iter = 2000,
  warmup = 1000,
  thin = 10,
  chains = 2,
  mean_cf = NA,
  var_cf = NA,
  centre_age = TRUE,
  ...
)
}
\arguments{
\item{input_data}{Dataframe with all variables}

\item{model_os}{Distribution name; exp, weibull, gompertz}

\item{model_pfs}{Distribution name; exp, weibull, gompertz}

\item{tx_name}{Treatment name; IPILIMUMAB, NIVOLUMAB, NIVOLUMAB+IPILIMUMAB}

\item{iter}{int}

\item{warmup}{Number of iterations burn-in; int}

\item{chains}{Number of chains; int}

\item{mean_cf}{Mean cure fraction; default to U\link{0,1}}

\item{var_cf}{Variance of cure fraction}

\item{centre_age}{Logical to centre regression covariate}

\item{...}{Additional arguments}
}
\description{
Use pre-compiled C code.
}
