% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmcm_joint_stan_file.R
\name{bmcm_joint_stan_file}
\alias{bmcm_joint_stan_file}
\title{Run Stan mixture cure joint model from Stan file}
\usage{
bmcm_joint_stan_file(
  input_data,
  model_os = "exp",
  model_pfs = "exp",
  tx_name = "IPILIMUMAB",
  iter = 2000,
  warmup = 1000,
  thin = 10,
  chains = 1,
  params_os = NA,
  params_pfs = NA,
  params_cf = list(mean_beta_cf = NA, var_beta_cf = NA),
  centre_age = TRUE,
  ...
)
}
\arguments{
\item{input_data}{Dataframe with all variables}

\item{model_os}{Distribution name; exp, weibull, gompertz}

\item{model_pfs}{Distribution name; exp, weibull, gompertz}

\item{tx_name}{Treatment name; IPILIMUMAB, NIVOLUMAB, NIVOLUMAB+IPILIMUMAB}

\item{iter}{integer}

\item{warmup}{Number of iterations burn-in; integer}

\item{chains}{Number of chains; integer}

\item{params_os}{List}

\item{params_pfs}{List}

\item{params_cf}{List of possible parameters values;
e.g. mean_beta_cf Mean cure fraction (default U\link{0,1})
var_beta_cf Variance of cure fraction}

\item{centre_age}{Logical to centre regression covariate}

\item{...}{Additional arguments}
}
\description{
Don't use pre-compiled C code.
}
