% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmcm_stan.R
\name{bmcm_stan}
\alias{bmcm_stan}
\title{bmcm_stan}
\usage{
bmcm_stan(
  input_data,
  formula,
  cureformula = ~1,
  family_latent = "exponential",
  prior_latent = NA,
  prior_cure = list(),
  centre_coefs = FALSE,
  bg_model = c("bg_distn", "bg_fixed"),
  bg_varname = "bg_rate",
  bg_hr = 1,
  t_max = 70,
  save_stan_code = FALSE,
  read_stan_code = FALSE,
  precompiled_model_path = NA,
  use_cmdstanr = FALSE,
  ...
)
}
\arguments{
\item{input_data}{A long-format data frame containing the input data.}

\item{formula}{An R formula object specifying the latent model component.}

\item{cureformula}{An R formula object specifying the cure fraction model component. Default is \code{~ 1}.}

\item{family_latent}{A string or vector specifying the distribution(s) for the latent model component.
Supported values are \code{"exp"}, \code{"weibull"}, \code{"gompertz"}, \code{"lognormal"}, \code{"loglogistic"}, and \code{"gengamma"}.}

\item{prior_latent}{An optional prior specification for the latent model component. Default is \code{NA}.}

\item{prior_cure}{An optional list specifying priors for the cure fraction model component. Default is an empty list.}

\item{centre_coefs}{Logical. If \code{TRUE}, coefficients will be centered. Default is \code{FALSE}.}

\item{bg_model}{A string specifying the background model type. Supported values are \code{"bg_distn"} and \code{"bg_fixed"}.}

\item{bg_varname}{A string specifying the background variable name in \code{input_data}. Default is \code{"bg_rate"}.}

\item{bg_hr}{A numeric value for the background hazard ratio adjustment. Default is \code{1}.}

\item{t_max}{A numeric value specifying the maximum time horizon for the model. Default is \code{70}.}

\item{save_stan_code}{Logical. If \code{TRUE}, saves the Stan model code to a file. Default is \code{FALSE}.}

\item{read_stan_code}{Logical. If \code{TRUE}, reads the Stan model code from a file instead of generating it dynamically. Default is \code{FALSE}.}

\item{precompiled_model_path}{A string specifying the path to a precompiled model. Default is \code{NA}.}

\item{use_cmdstanr}{Logical. If \code{TRUE}, the \code{cmdstanr} package is used to fit the model. Default is \code{FALSE}.}

\item{...}{Additional parameters to pass to the Stan sampler.}
}
\value{
A list of class \code{bmcm} containing the Stan model output, MCMC parameters,
function call, distributions used, model inputs, input data, and formulas for the cure and latent components.
}
\description{
Jointly estimates all treatments and generates Stan code for Bayesian relative mixture cure modeling.
It supports various latent and cure model configurations and allows the use of precompiled models or on-the-fly compilation.
}
\examples{
\dontrun{
data("surv_input_data", package = "multimcm")
out <- bmcm_stan(
  input_data = surv_input_data,
  formula = "Surv(time=os, event=os_event) ~ 1",
  cureformula = "~ TRTA + (1 | center_id)",
  family_latent = "exponential",
  bg_model = "bg_fixed",
  bg_varname = "rate",
  t_max = 400
)
}

}
