% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmcm_stan_file.R
\name{bmcm_stan_file}
\alias{bmcm_stan_file}
\title{Run Stan mixture cure model from Stan file}
\usage{
bmcm_stan_file(
  input_data,
  model = "exp",
  event_type = "PFS",
  tx_name = "IPILIMUMAB",
  iter = 3000,
  warmup = 1000,
  thin = 10,
  chains = 1,
  mean_cf = NA,
  var_cf = NA,
  centre_age = TRUE,
  ...
)
}
\arguments{
\item{input_data}{Dataframe with all variables}

\item{model}{Distribution name; exp, weibull, gompertz}

\item{event_type}{Overall survival or progression-free survival; OS, PFS}

\item{tx_name}{Treatment name; IPILIMUMAB, NIVOLUMAB, NIVOLUMAB+IPILIMUMAB}

\item{iter}{int}

\item{warmup}{Number of iterations burn-in; int}

\item{chains}{Number of chains; int}

\item{mean_cf}{Mean cure fraction; default to U\link{0,1}}

\item{var_cf}{Variance of cure fraction}

\item{centre_age}{Logical to centre regression covariate}

\item{...}{Additional arguments}
}
\description{
Does not use pre-compiled C code.
}
