% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_forest_cutpoint.R
\name{cf_forest_cutpoint}
\alias{cf_forest_cutpoint}
\title{Cure fraction forest plot using cut-point Stan output}
\usage{
cf_forest_cutpoint(
  distns = list(c("exp", "exp"), c("lognormal", "lognormal")),
  folder = NA,
  save_name = c("_12", "_30", "_100"),
  is_hier = TRUE
)
}
\arguments{
\item{distns}{List of distributions}

\item{folder}{String of location to read in data from}

\item{save_name}{File name appended text}

\item{is_hier}{Is hierarchical model (or separate)? i.e. whether to use a global parameter}
}
\value{
A ggplot object
}
\description{
Using all treatment Stan model.
}
\examples{
\dontrun{
 # independent model
 fp_sep <- cf_forest_cutpoint(folder = "data/dbl_cut/separate", save_name = c("_30", "_12", "_100"))
 #ggsave(fp_sep, filename = "plots/forest_plot_cf_sep_cpt.png",
 #       dpi = 640, width = 16, height = 14)

 # hierarchical model
 fp_hier <- cf_forest_cutpoint(folder = "data/dbl_cut/hier", save_name = c("_30", "_12", "_100"))
 #ggsave(fp_hier, filename = "plots/forest_plot_cf_hier_cpt.png",
 #       dpi = 640, width = 16, height = 14)
}
}
