% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_forest_plot.R
\name{cf_forest_plot}
\alias{cf_forest_plot}
\title{Cure fraction forest plot using Stan output}
\usage{
cf_forest_plot(folder = "data/")
}
\arguments{
\item{folder}{A string specifying the directory path containing the saved Stan model output files.}
}
\value{
A \code{ggplot2} object showing the cure fraction forest plot.
}
\description{
Generates a forest plot displaying the estimated cure fractions across all treatments
derived from the joint relative survival mixture cure model output.
}
\examples{
\dontrun{
 # independent model
 fp_sep <- cf_forest_plot("data/separate/")
 #ggsave(fp_sep, filename = "plots/forest_plot_sep_multimcm.png", dpi = 640, width = 12, height = 8)

 # hierarchical model
 fp_hier <- cf_forest_plot("data/")
 #ggsave(fp_hier, filename = "plots/forest_plot_multimcm.png", dpi = 640, width = 12, height = 8)
}
}
