% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stancode_postpred.R
\name{create_stancode_postpred}
\alias{create_stancode_postpred}
\title{Posterior prediction Stan code builder}
\usage{
create_stancode_postpred(
  models = c("exp", "weibull", "gompertz", "loglogistic", "lognormal")
)
}
\description{
Posterior prediction Stan code builder
}
\keyword{internal}
