% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots.R
\name{curefraction_boxplots}
\alias{curefraction_boxplots}
\title{curefraction_boxplots}
\usage{
curefraction_boxplots(stan_out)
}
\arguments{
\item{stan_out}{stan model output}
}
\description{
do for OS, PFS
}
