% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_prior_cure.R
\name{default_prior_cure}
\alias{default_prior_cure}
\title{Prepare shared parameters for Stan model}
\usage{
default_prior_cure(formula_cure, bg_model = 2)
}
\arguments{
\item{formula_cure}{Result of \code{parse_formula}}

\item{bg_model}{Background model index:
\enumerate{
\item exponential distribution
\item fixed point values from life-tables
}}
}
\value{
list
}
\description{
Model parameters for cure fraction, background and generated.
}
