% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_arrange_shared_legend.R
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{Use single legend for grid of plots}
\usage{
grid_arrange_shared_legend(
  ...,
  ncol = length(list(...)),
  nrow = 1,
  position = c("bottom", "right", "top")
)
}
\arguments{
\item{...}{ggplot objects}

\item{ncol}{number of columns}

\item{nrow}{number of rows}

\item{position}{position of the legend}
}
\description{
from https://cran.r-project.org/web/packages/egg/vignettes/Ecosystem.html
}
