% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_code_blocks.R
\name{make_latent_model_code}
\alias{make_latent_model_code}
\title{Make_latent_model_code}
\usage{
make_latent_model_code(model, id = 1L)
}
\description{
Make_latent_model_code
}
\examples{
\dontrun{
lapply(make_latent_model_code("lognormal"), cat)
}

}
\keyword{internal}
