% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoM_beta.R
\name{MoM_beta}
\alias{MoM_beta}
\title{MoM_beta}
\usage{
MoM_beta(xbar, vbar)
}
\arguments{
\item{xbar}{Mean}

\item{vbar}{Variance}
}
\value{
Beta distribution parameters a, b in a list.
}
\description{
Method of moments for the Beta distribution.
}
\examples{

}
