% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_prior_latent.R
\name{pars_format}
\alias{pars_format}
\title{ensure consistent dimensions
depending on number of covariates}
\usage{
pars_format(nvars)
}
\description{
ensure consistent dimensions
depending on number of covariates
}
