% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{parse_formula}
\alias{parse_formula}
\title{Parse formula}
\usage{
parse_formula(formula, data, family = NA)
}
\arguments{
\item{formula}{A formula object}

\item{data}{A data frame}

\item{family}{Family string}
}
\description{
Parse formula
}
\keyword{internal}
