% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_S_event_type.R
\name{plot_S_event_type}
\alias{plot_S_event_type}
\title{plot_S_event_type}
\usage{
plot_S_event_type(file_names)
}
\arguments{
\item{file_names}{Nested list of file name for stan output}
}
\value{
ggplot object
}
\description{
Plot results of running Stan
relative survival mixture cure model.
}
\examples{
load("data/file_names.RData")

}
