% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_S_joint_data.R
\name{prep_S_joint_data.default}
\alias{prep_S_joint_data.default}
\title{already extracted matrix}
\usage{
\method{prep_S_joint_data}{default}(bmcm_out, ...)
}
\arguments{
\item{bmcm_out}{Output of Stan model}

\item{...}{Additional arguments}
}
\description{
already extracted matrix
}
