% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_S_joint_data.R
\name{prep_S_joint_data}
\alias{prep_S_joint_data}
\title{Prepare data for survival plot}
\usage{
prep_S_joint_data(bmcm_out, ...)
}
\arguments{
\item{bmcm_out}{Output of Stan model}

\item{...}{Additional arguments}
}
\description{
Prepare data for survival plot
}
