% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_stan_data.R
\name{prep_stan_data}
\alias{prep_stan_data}
\title{prep_stan_data}
\usage{
prep_stan_data(input_data, event_type, tx_name, centre_age)
}
\arguments{
\item{input_data}{surv_input_data}

\item{event_type}{PFS, OS}

\item{tx_name}{IPILIMUMAB, NIVOLUMAB, NIVOLUMAB+IPILIMUMAB}

\item{centre_age}{Logical}
}
\value{
List;
sample size,
times,
censoring indicator,
number of covariates,
covariates
}
\description{
Data specific to OS or PFS for Stan input.
}
