% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurv.R
\name{rsurv}
\alias{rsurv}
\title{Generate time to event sample with censoring}
\usage{
rsurv(
  n = 100,
  distn = "exp",
  prop_cens = 0.1,
  params = list(mu = c(-5, 0.005)),
  X = rep(c(10, 25, 50, 100), each = 25)
)
}
\description{
Generate time to event sample with censoring
}
\examples{

rsurv(n = 100,
      distn = "exp",
      prop_cens = 0.1,
      params = list(mu = c(-5, 0.005)),
      X = rep(c(10, 25, 50, 100), each = 25))

}
