% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurv.R
\name{rsurv_mix}
\alias{rsurv_mix}
\title{Survival mixture model simulation}
\usage{
rsurv_mix(
  cf = 0.2,
  n = 200,
  distn = c("exp", "exp"),
  prop_cens = 0.1,
  params = list(list(mu = c(2.5, 0.005)), list(mu = c(-8, 0.005))),
  X = rep(c(10, 25, 50, 100), each = 50)
)
}
\description{
Survival mixture model simulation
}
\examples{

rsurv_mix(cf = 0.2,
          n = 200,
          distn = c("exp", "exp"),
          prop_cens = 0.1,
          params =
            list(
              list(
                mu = c(2.5, 0.005)),
              list(
                mu = c(-8, 0.005))),
          X = rep(c(10, 25, 50, 100), each = 50))

}
