% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmcm_stan.R
\name{bmcm_stan}
\alias{bmcm_stan}
\title{bmcm_stan}
\usage{
bmcm_stan(
  input_data,
  formula,
  cureformula = ~1,
  family_latent = "exponential",
  prior_latent = NA,
  prior_cure = NA,
  centre_coefs = FALSE,
  bg_model = c("bg_distn", "bg_fixed"),
  bg_varname = "bg_rate",
  bg_hr = 1,
  t_max = 60,
  ...
)
}
\arguments{
\item{input_data}{Long format data frame}

\item{formula}{R formula object for latent model component}

\item{cureformula}{R formula object for cure fraction model component}

\item{family_latent}{Name of distribution, from
"exp", "weibull", "gompertz", "lognormal", "gamma".
Single string or vector of length number of end-points}

\item{prior_latent}{Optional}

\item{prior_cure}{Optional}

\item{centre_coefs}{Logical}

\item{bg_model}{User supplied distribution or estimated.
In future this will be probabilities or fitted model object.}

\item{bg_varname}{Background variable name in \code{input_data}}

\item{bg_hr}{Background hazard ratio adjustment}

\item{t_max}{Maximum time horizon}

\item{...}{Additional parameters}
}
\value{
Stan output as \code{bmcm} class
}
\description{
jointly estimate all treatments
generate Stan code
}
