% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stancode.R
\name{create_stancode}
\alias{create_stancode}
\title{Create Stan code from component parts}
\usage{
create_stancode(models)
}
\arguments{
\item{models}{Vector survival model names}
}
\value{
string
}
\description{
Create Stan code from component parts
}
\examples{
cat(create_stancode(c("exp", "exp")))

\dontrun{
 check with
 writeLines(create_stancode(c("gengamma", "exp")), "temp.stan")
}
}
