% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_kaplan_meier.R
\name{geom_kaplan_meier}
\alias{geom_kaplan_meier}
\title{Geom for Kaplan-Meier ggplot}
\usage{
geom_kaplan_meier(out_dat, col = "black", add_marks = TRUE)
}
\description{
Geom for Kaplan-Meier ggplot
}
