% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_block_code.R
\name{make_latent_model_code}
\alias{make_latent_model_code}
\title{make_latent_model_code}
\usage{
make_latent_model_code(model, id = 1L)
}
\description{
make_latent_model_code
}
\examples{
lapply(make_latent_model_code("lognormal"), cat)

}
