% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_S_joint.R
\name{plot_S_joint}
\alias{plot_S_joint}
\title{plot_S_joint}
\usage{
plot_S_joint(file_names = NA, stan_list = NA)
}
\arguments{
\item{file_names}{Nested list of file names for Stan output}

\item{stan_list}{List of Stan output}
}
\value{
ggplot object
}
\description{
Plot results of running Stan
relative survival joint mixture cure model.
}
\examples{
load("data/file_names.RData")

}
