% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_S_joint.R
\name{plot_S_joint}
\alias{plot_S_joint}
\title{Plot survival curves for joint model and all treatments}
\usage{
plot_S_joint(
  bmcm_out,
  facet = TRUE,
  annot_cf = FALSE,
  add_km = FALSE,
  add_marks = TRUE,
  ...
)
}
\arguments{
\item{facet}{Two separate plots for each end point or overlaid?}

\item{annot_cf}{Annotate with cure fractions? Logical}

\item{add_km}{Include Kaplan-Meier layer? Logical}

\item{add_marks}{Kaplan-Meier censoring marks; logical}

\item{...}{Additional parameters}

\item{out}{bmcm class output list}
}
\value{
ggplot object
}
\description{
Use results of running Stan with \code{bmcm_stan()}
relative survival joint mixture cure model.
}
\examples{
load("data/file_names.RData")

}
