% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precompile_bmcm_model.R
\name{precompile_bmcm_model}
\alias{precompile_bmcm_model}
\title{Precompile bmcm model}
\usage{
precompile_bmcm_model(
  input_data,
  family_latent = "exponential",
  cureformula = ~1,
  model_name = NA,
  use_cmdstanr = FALSE,
  file_path = NA
)
}
\arguments{
\item{input_data}{Data frame}

\item{family_latent}{String specifying latent distribution family}

\item{cureformula}{Formula for cure fraction}

\item{model_name}{Optional string for model name}

\item{use_cmdstanr}{Logical. If TRUE, use cmdstanr to compile the model. Default is FALSE.}

\item{file_path}{Optional string for file path to save compiled model}
}
\value{
A compiled stan model object
}
\description{
Precompile bmcm model
}
