% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precompile_bmcm_model.R
\name{precompile_bmcm_model}
\alias{precompile_bmcm_model}
\title{Precompile bmcm model}
\usage{
precompile_bmcm_model(
  input_data,
  family_latent = "exponential",
  cureformula = ~1,
  model_name = NA,
  use_cmdstanr = FALSE
)
}
\arguments{
\item{use_cmdstanr}{Logical. If TRUE, use cmdstanr to compile the model. Default is FALSE.}
}
\description{
Precompile bmcm model
}
