% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_bg_data.R
\name{prep_bg_data}
\alias{prep_bg_data}
\title{Prepare background data}
\usage{
prep_bg_data(input_data, bg_varname, formula_cure, event_type, suffix = TRUE)
}
\arguments{
\item{suffix}{}
}
\description{
Background hazard point values.
}
