% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_latent_data.R
\name{prep_latent_data}
\alias{prep_latent_data}
\title{Prepare Stan data in latent model}
\usage{
prep_latent_data(
  input_data,
  formula_cure,
  formula_latent,
  event_type,
  centre_coefs = FALSE,
  suffix = TRUE
)
}
\arguments{
\item{formula_cure}{parsed formula}

\item{formula_latent}{parsed formula}

\item{event_type}{cluster/group}

\item{centre_coefs}{Logical
1: Exponential distribution; 2: fixed point values from life-table}
}
\value{
List consisting of:
sample size,
times,
censoring indicator,
number of covariates,
covariates
}
\description{
Data specific to end type for Stan input.
}
