% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_S_joint_data.R
\name{prep_S_joint_data.bmcm}
\alias{prep_S_joint_data.bmcm}
\title{Prepare data for survival plot}
\usage{
\method{prep_S_joint_data}{bmcm}(bmcm_out)
}
\arguments{
\item{bmcm_out}{Output of Stan model}
}
\value{
data frame
}
\description{
Prepare data for survival plot
}
