% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_shared_params.R
\name{prep_shared_params}
\alias{prep_shared_params}
\title{prep_shared_params}
\usage{
prep_shared_params(
  mean_beta_cf = NA,
  var_beta_cf = NA,
  mu_cf = NA,
  sigma_cf = NA,
  mu_bg = c(-8.5, 0.03),
  sigma_bg = c(1, 1),
  t_max = 60,
  mu_joint = 0,
  sigma_joint = 0.1
)
}
\arguments{
\item{mean_beta_cf}{Mean of cure fraction to use in method of
moments (MoM) to obtain Beta distribution
hyper-parameters a, b}

\item{var_beta_cf}{Variance of cure fraction for MoM}

\item{mu_cf}{Mean of cure fraction to use directly}

\item{sigma_cf}{Standard deviation of cure fraction to use directly}

\item{mu_bg}{Mean of background survival}

\item{sigma_bg}{Standard deviation of background survival}

\item{t_max}{Time horizon}

\item{sigma_joint}{}
}
\value{
list
}
\description{
Model parameters not specific to OS or PFS:
cure fraction, background and generated.
}
