% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_stan_data.R
\name{prep_stan_data}
\alias{prep_stan_data}
\title{prep_stan_data}
\usage{
prep_stan_data(input_data, event_type, tx_name, centre_age)
}
\arguments{
\item{centre_age}{}
}
\value{
List
}
\description{
Data specific to OS or PFS for stan input.
}
