% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_prior_latent.R
\name{prep_stan_params}
\alias{prep_stan_params}
\alias{default_prior_latent}
\title{default_prior_latent}
\usage{
default_prior_latent(formula_latent, formula_cure)
}
\value{
hyper-parameters
}
\description{
Default parameters for priors if not supplied.
}
