% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_stan_params.R
\name{prep_stan_params}
\alias{prep_stan_params}
\title{prep_stan_params}
\usage{
prep_stan_params(model, params = NA)
}
\arguments{
\item{model}{exp, weibull}
}
\value{
hyper-parameters
}
\description{
prep_stan_params
}
